// workshop-requisition-codeapp/src/generated/models/Wksp_approvalhistoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Wksp_approvalhistoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Wksp_approvalhistoriesstatecode = keyof typeof Wksp_approvalhistoriesstatecode;
export const Wksp_approvalhistoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Wksp_approvalhistoriesstatuscode = keyof typeof Wksp_approvalhistoriesstatuscode;
export const Wksp_approvalhistorieswksp_action = {
  100000000: 'Approved',
  100000001: 'Rejected'
} as const;
export type Wksp_approvalhistorieswksp_action = keyof typeof Wksp_approvalhistorieswksp_action;
export const Wksp_approvalhistorieswksp_approvallevel = {
  100000000: 'Supervisor',
  100000001: 'Manager',
  100000002: 'Finance'
} as const;
export type Wksp_approvalhistorieswksp_approvallevel = keyof typeof Wksp_approvalhistorieswksp_approvallevel;

export interface Wksp_approvalhistoriesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Wksp_approvalhistoriesstatecode;
  statuscode?: Wksp_approvalhistoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  wksp_action: Wksp_approvalhistorieswksp_action;
  wksp_approvalhistoryid: string;
  wksp_approvallevel: Wksp_approvalhistorieswksp_approvallevel;
  wksp_approvalnumber?: string;
  "wksp_ApproverId@odata.bind"?: string;
  wksp_comment?: string;
  wksp_decisiondate: string;
  "wksp_RequisitionId@odata.bind"?: string;
}

export interface Wksp_approvalhistories extends Wksp_approvalhistoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  wksp_actionname?: string;
  wksp_approvallevelname?: string;
  wksp_approveridname?: string;
  wksp_approveridyominame?: string;
  wksp_requisitionidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  wksp_approverid?: object;
  _wksp_approverid_value?: string;
  wksp_requisitionid?: object;
  _wksp_requisitionid_value?: string;
}